added courses to mongo

+++ src/main/java/com/project/dreamteam/repository/UserRepository.java
@@ -1,5 +1,5 @@
 package com.project.dreamteam.repository;
 import com.project.dreamteam.model.User;
 import org.springframework.data.mongodb.repository.*;
-public interface UserRepository extends MongoRepository<User, Integer> {
+public interface UserRepository extends MongoRepository<User, Long> {
 }
